Commit fixes the upgrade issue with KAPEX
This change fixes an illegal code bug, that occurred after a line was mistakenly uncommented. The impact of this bug is that KAPEX cannot be upgraded.

The commit that was used to generate the genesis state and the WASM code was #b609742c0f394a55712a0a883286b38232600b67

This correction is a fork of that commit, with the offending line re-commented.

A code substitute set of states has been generated for Lego (local Rococo) and Kapex (Production)

This fix will be merged with Main in order to produce the code substitute genesis and wasm code for Stagex (Public Rococo)